Fix float16/float32 dtype mismatch in retrieve_faiss_sparse
Corpus embeddings are now stored as float16 (OOM fix from prev commit).
retrieve_faiss_sparse slices them and passes to torch.mm, which requires
both operands to be the same dtype. Cast both q_norm and c_norm to
float32 explicitly before mm.

+++ scripts/eval_beir.py
@@ -254,8 +254,8 @@ def retrieve_faiss_sparse(query_embs: np.ndarray, corpus_embs: np.ndarray,
                 np.ascontiguousarray(q_norm.astype(np.float32)), k
             )
         else:
-            q_t = torch.from_numpy(q_norm)
-            c_t = torch.from_numpy(c_norm)
+            q_t = torch.from_numpy(q_norm.astype(np.float32))
+            c_t = torch.from_numpy(c_norm.astype(np.float32))
             sim = torch.mm(q_t, c_t.t())
             scores_t, indices_t = sim.topk(min(k, N_c), dim=-1)
             scores = scores_t.numpy()
